perf(ui): load preferences when settings opens
