+++ apps/ui/playwright.config.ts
@@ -113,7 +113,7 @@ export default defineConfig({
       // peak.
       name: "interaction",
       testMatch:
-        /(about-state|account-detail-state|active-entity|agents-state|api-account-semantics|api-directory-state|block-detail|block-stream-prefetch|chain-overview-state|chain-stream-state|charts|command-palette-preload|compare-loading|content-runtime-isolation|contribute-state|deferred-stake-flow|directory-secondary-state|docs-mcp-layout|event-detail|extrinsic-detail-state|footer-freshness-state|health-secondary-state|home-secondary-state|live-block-rail-prefetch|provider-detail-state|providers-index-state|rank|revenue-state|schemas-state|settings-state|subnet-secondary-state|validator-detail-state|crawlable-subnet-index|evidence-deep-link|indexable-routes|keyboard|multisig-related-error|offline|rendered-reader-copy|sticky-table-header|validator-unit-boundaries)\.spec\.ts$/,
+        /(about-state|account-detail-state|active-entity|agents-state|api-account-semantics|api-directory-state|block-detail|block-stream-prefetch|chain-overview-state|chain-stream-state|charts|command-palette-preload|compare-loading|content-runtime-isolation|contribute-state|deferred-stake-flow|directory-secondary-state|docs-mcp-layout|event-detail|extrinsic-detail-state|footer-freshness-state|health-secondary-state|home-secondary-state|live-block-rail-prefetch|preferences-loading|provider-detail-state|providers-index-state|rank|revenue-state|schemas-state|settings-state|subnet-secondary-state|validator-detail-state|crawlable-subnet-index|evidence-deep-link|indexable-routes|keyboard|multisig-related-error|offline|rendered-reader-copy|sticky-table-header|validator-unit-boundaries)\.spec\.ts$/,
       dependencies: ["overflow"],
       // Serial within the phase costs a few seconds and removes the last
       // source of self-contention for exactly the tests that proved sensitive
